+++ modules/util/portal-tools-sample-sql-builder/src/main/java/com/liferay/portal/tools/sample/sql/builder/SampleSQLBuilder.java
@@ -252,6 +252,8 @@ public class SampleSQLBuilder {
 							"csvFileWriter", csvFileWriter
 						).put(
 							"dataFactory", new DataFactory()
+						).put(
+							"journalDataFactory", new JournalDataFactory()
 						).build(),
 						sampleSQLWriter);
 				}
